Add NavigationIdEndpointAnalyzer for FastEndpoints
Introduce a new diagnostic analyzer, `NavigationIdEndpointAnalyzer`, enforcing that all FastEndpoints endpoints implement the `IHasNavigationId` interface (rule `DSMPS010`). Update the `DummyNavigationData` method in `NavigationServiceTest.cs` to change the parameter type for `BuildTree` from `int?` to `short?`. Document the new rule in `AnalyzerReleases.Unshipped.md`.

+++ src/services/tests/Demoulas.ProfitSharing.UnitTests/Endpoints/Navigations/NavigationServiceTest.cs
@@ -97,8 +97,8 @@ public class NavigationServiceTests : ApiTestBase<Program>
     //Dummy Data
     private List<NavigationDto> DummyNavigationData()
     {
-        var lookup = _navigationListObj.ToLookup(x => x.ParentId);
-        List<NavigationDto> BuildTree(int? parentId)
+    var lookup = _navigationListObj.ToLookup(x => x.ParentId);
+    List<NavigationDto> BuildTree(short? parentId)
         {
             return lookup[parentId]
                 .Select(x => new NavigationDto
